Refactor SDL_events.h and SDL_render.h for improved clarity and maintainability
- Replaced extensive copyright comments in SDL_events.h and SDL_render.h with concise headers to eliminate copyright dependencies.
- Consolidated event type enumerations and structures in SDL_events.h for better organization and readability.
- Updated SDL_render.h to streamline rendering-related structures and removed unnecessary comments, enhancing overall usability.
- Adjusted SDLHelpers.swift to change texture dimensions from Float to Int32 for consistency with SDL_GetTextureSize, improving type safety.

+++ Sources/Tetrix/SDLHelpers.swift
@@ -492,11 +492,11 @@ class Texture {
         guard let texture = SDL_CreateTextureFromSurface(renderer, surface) else { return nil }
         SDL_DestroySurface(surface)
         
-        var width: Float = 0
-        var height: Float = 0
+        var width: Int32 = 0
+        var height: Int32 = 0
         SDL_GetTextureSize(texture, &width, &height)
         
-        return Texture(sdlTexture: texture, width: width, height: height)
+        return Texture(sdlTexture: texture, width: Float(width), height: Float(height))
     }
     
     /// Create texture from text (Linux only - uses SDL3 text rendering)
